remove(page): map/page.tsx, replace by page.tsx in /
