+++ app/robots.ts
@@ -5,7 +5,7 @@ export default function robots(): MetadataRoute.Robots {
     rules: [
       {
         userAgent: "*",
-        allow: ["/", "/map"],
+        allow: ["/"],
         disallow: "/debug",
       },
     ],
